move wcs code into sub package
